feat: add :core-designsystem, :core-model, :core-common, :core-domain (NIA slice 2)
Introduces the canonical NIA module split so adopters know where each
kind of code belongs.

New modules:
- :core-designsystem (Android library + compose) — owns ConsultMeTheme,
  Color, Type. Theme moved out of :app/ui/theme/ into
  com.thecompany.consultme.core.designsystem.theme. BREAKING for any
  fork that imported the old theme path.
- :core-model (jvm-library) — pure-Kotlin scaffold for domain data
  classes. No Android, no Hilt.
- :core-common (jvm-library) — pure-Kotlin shared utilities. Ships NIA's
  Dispatcher qualifier + AppDispatchers enum. Adopters provide @Provides
  bindings wherever they apply Hilt.
- :core-domain (jvm-library) — pure-Kotlin scaffold for use-cases.
  Depends on :core-model.

Convention plugin update:
- consultme.android.feature now adds :core-designsystem and :core-ui as
  implementation deps, mirroring NIA's AndroidFeatureConventionPlugin.
  Feature modules get the theme + shared composables for free.

Catalog additions (consumed by the new pure-Kotlin modules):
- javax-inject (for @Qualifier in :core-common)
- kotlinx-coroutines-core (JVM-only variant; -android is for Android
  modules, -test for tests)

Also pins AGP 8 default behaviors in gradle.properties before the AGP 9
migration in Phase 9.

Refs #122 (NIA-alignment tracking).
