+++ app/src/main/java/com/thecompany/consultme/MainActivity.kt
@@ -11,9 +11,9 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
+import com.thecompany.consultme.core.designsystem.theme.ConsultMeTheme
 import com.thecompany.consultme.feature.example.ui.ExampleScreen
 import com.thecompany.consultme.feature.example.ui.ExampleUiState
-import com.thecompany.consultme.ui.theme.ConsultMeTheme
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
