+++ build-logic/convention/src/main/kotlin/consultme.android.feature.gradle.kts
@@ -10,6 +10,9 @@ plugins {
 val libs = the<LibrariesForLibs>()
 
 dependencies {
+    "implementation"(project(":core-designsystem"))
+    "implementation"(project(":core-ui"))
+
     "implementation"(libs.androidx.lifecycle.runtime.compose)
     "implementation"(libs.androidx.lifecycle.viewmodel.compose)
     "implementation"(libs.androidx.hilt.navigation.compose)
