+++ core-common/build.gradle.kts
@@ -1,0 +1,9 @@
+// Copyright 2025 MyCompany
+plugins {
+    id("consultme.jvm.library")
+}
+
+dependencies {
+    implementation(libs.javax.inject)
+    implementation(libs.kotlinx.coroutines.core)
+}
