+++ core-common/src/main/kotlin/com/thecompany/consultme/core/common/dispatchers/AppDispatchers.kt
@@ -1,0 +1,7 @@
+// Copyright 2025 MyCompany
+package com.thecompany.consultme.core.common.dispatchers
+
+enum class AppDispatchers {
+    Default,
+    IO,
+}
