+++ core-common/src/main/kotlin/com/thecompany/consultme/core/common/dispatchers/Dispatcher.kt
@@ -1,0 +1,8 @@
+// Copyright 2025 MyCompany
+package com.thecompany.consultme.core.common.dispatchers
+
+import javax.inject.Qualifier
+
+@Qualifier
+@Retention(AnnotationRetention.RUNTIME)
+annotation class Dispatcher(val dispatcher: AppDispatchers)
