+++ core-designsystem/src/main/java/com/thecompany/consultme/core/designsystem/theme/Theme.kt
@@ -1,5 +1,5 @@
 // Copyright 2025 MyCompany
-package com.thecompany.consultme.ui.theme
+package com.thecompany.consultme.core.designsystem.theme
 
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
@@ -21,16 +21,6 @@ private val LightColorScheme = lightColorScheme(
     primary = Purple40,
     secondary = PurpleGrey40,
     tertiary = Pink40,
-
-    /* Other default colors to override
-    background = Color(0xFFFFFBFE),
-    surface = Color(0xFFFFFBFE),
-    onPrimary = Color.White,
-    onSecondary = Color.White,
-    onTertiary = Color.White,
-    onBackground = Color(0xFF1C1B1F),
-    onSurface = Color(0xFF1C1B1F),
-     */
 )
 
 @Composable
