+++ core-domain/build.gradle.kts
@@ -1,0 +1,8 @@
+// Copyright 2025 MyCompany
+plugins {
+    id("consultme.jvm.library")
+}
+
+dependencies {
+    implementation(projects.coreModel)
+}
