+++ core-model/build.gradle.kts
@@ -1,0 +1,4 @@
+// Copyright 2025 MyCompany
+plugins {
+    id("consultme.jvm.library")
+}
