+++ gradle.properties
@@ -35,3 +35,13 @@ org.gradle.caching=true
 # as $YEAR so spotlessApply rewrites the current year on demand.
 template.company=MyCompany
 template.licenseYear=$YEAR
+android.defaults.buildfeatures.resvalues=true
+android.sdk.defaultTargetSdkToCompileSdkIfUnset=false
+android.enableAppCompileTimeRClass=false
+android.usesSdkInManifest.disallowed=false
+android.uniquePackageNames=false
+android.dependency.useConstraints=true
+android.r8.strictFullModeForKeepRules=false
+android.r8.optimizedResourceShrinking=false
+android.builtInKotlin=false
+android.newDsl=false
